Alpha 0.06
A revamp of the way Entitys are organized in preparation for future features and changes to the Movement system to deal with the former odd behaviors of entity movement.

EntityCollection (new)
-----------------------
An entity that is a collection of other unique Entitys, including childern EntityCollections. Checks for uniqueness in members and  EntityMembers are checked for unique coordinates. See EntityMembers. Child EntityCollections are not checked for unique coordinates as they do not represent a single Entity, so may span across many coordinates.

EntityMember (new)
----------------------
All Entitys that are not EntityCollections are subclasses of this. They represent a single Entity.

Movable & Movement (new)
-------------------------------
The component on entity that holds data for their movements on screen is now Movable. Movement is now a data type holding data for a single movement action.

MovementManager (changes)
--------------------------------
Still manages the screen movements of Entities, but is updated to use Movable components and the Movement data type. Also executes movement events for the new IMovementEventHandler interface

IMovementEventHandler (new)
---------------------------------
First implementation of an event handler. An interface managers can implement to handle movement events executed by the MovementManager. Includes OnMovementCreate(Movement), OnMovementChange(Movement), OnMovementStart(Movement), OnMovementEnd(Movement)